requirements.txt
chromadb
sentence-transformers
pydantic-settings 
pydantic<2.12
fastapi
uvicorn
openai
python-dotenv
requests
mcp